Move plugin back out of operator
